requirements.txt
asgiref==3.7.2
boto3==1.28.77
botocore==1.31.77
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.5
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
